ovdet/requirements.txt
absl-py==2.1.0
accelerate==0.32.1
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
anytree==2.12.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
black==21.4b2
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
blis==0.7.11
boto3==1.34.140
botocore==1.34.140
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
colorama==0.4.6
confection==0.1.5
contexttimer==0.3.3
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses==0.6
datasets==2.21.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.7.0
# detectron2==0.6
dill==0.3.7
distlib==0.3.8
distro==1.9.0
einops==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.4
fasttext==0.9.3
filelock==3.14.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.6.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.64.1
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hydra-core==1.1.2
hyperframe==6.0.1
icecream==2.1.3
identify==2.6.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib-resources==5.2.3
inflect==7.3.1
iopath==0.1.9
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==1.6.17
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lvis==0.5.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==2.0.1
mmcv-full==1.7.2
mmdet==3.0.0
mmengine==0.10.4
model-index==0.1.11
modelscope==1.16.0
more-itertools==10.3.0
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.6.2
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
olefile==0.47
omegaconf==2.1.2
openai==1.35.10
opencv-python==4.6.0.66
opencv-python-headless==4.5.5.64
opendatalab==0.0.10
opendatasets==0.1.22
openmim==0.3.9
# openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
peft==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.24.0
portalocker==2.10.0
pre-commit==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydot==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.4.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
segment_anything==1.0
sentencepiece==0.2.0
setuptools==60.2.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.38.0
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
# torch==1.13.1+cu116
# torchaudio==0.13.1+cu116
# torchvision==0.14.1+cu116
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.26.0
triton==2.3.1
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.57
ultralytics-thop==2.0.0
urllib3==1.26.19
virtualenv==20.26.4
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
zstandard==0.19.0
